git-svn-id: https://svn2.hosted-projects.com/jjinux/jenarix@69 af7818e5-df21-49d5-b9e6-ecf3e1c3d4f7

+++ sandbox/warren/090514/src/json/jx_json_out.c
@@ -248,7 +248,7 @@ jx_ob jx_ob_to_json_with_flags(jx_ob ob, jx_int flags)
       return jx_ob_from_str(buffer);
     }
     break;
-  case JX_META_BIT_IDENT: /* Jenarix JSON extension */
+  case JX_META_BIT_IDENT: /* Jenarix identifiers (extended JSON) */
     if(ob.meta.bits & JX_META_BIT_GC) {
       return jx_ob_from_str(ob.data.io.str);
     } else {
@@ -261,7 +261,7 @@ jx_ob jx_ob_to_json_with_flags(jx_ob ob, jx_int flags)
       if(ob.meta.bits & JX_META_BIT_GC) {
         sprintf(buffer,"*builtin_%p*",(void*)ob.data.io.builtin);
       } else {
-        sprintf(buffer,"*builtin_%03d*",ob.data.io.int_);
+        sprintf(buffer,"*builtin_%03d*",(int)ob.data.io.int_);
       }
       return jx_ob_from_str(buffer);
     }
