+++ sandbox/warren/090514/src/test/test001.c
@@ -44,29 +44,29 @@ int main(int argc, char **argv)
 
     P1("1 && %p", (void *) vla);
     P1("0 == %d", jx_vla_insert(&vla, 0, 1));
-    P1("1 == %d", jx_vla_size(&vla));
+    P1("1 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_insert(&vla, 1, 1));
     P1("0 == %d", jx_vla_insert(&vla, 0, 1));
-    P1("3 == %d", jx_vla_size(&vla));
+    P1("3 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_free(&vla));
   }
 
   {
     jx_int *vla = jx_vla_new(sizeof(jx_int), 10);
     P1("1 && %p", (void *) vla);
-    P1("10 == %d", jx_vla_size(&vla));
+    P1("10 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_insert(&vla, 0, 10));
-    P1("20 == %d", jx_vla_size(&vla));
+    P1("20 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_remove(&vla, 5, 10));
-    P1("10 == %d", jx_vla_size(&vla));
+    P1("10 == %d", (int)jx_vla_size(&vla));
     P1("-1 == %d", jx_vla_remove(&vla, 5, 10));
-    P1("10 == %d", jx_vla_size(&vla));
+    P1("10 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_append(&vla, 5));
-    P1("15 == %d", jx_vla_size(&vla));
+    P1("15 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_remove(&vla, 5, 10));
-    P1("5 == %d", jx_vla_size(&vla));
+    P1("5 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_remove(&vla, 0, 5));
-    P1("0 == %d", jx_vla_size(&vla));
+    P1("0 == %d", (int)jx_vla_size(&vla));
     P1("0 == %d", jx_vla_free(&vla));
   }
