+++ sandbox/warren/090514/src/test/test003.c
@@ -50,19 +50,19 @@ int main(int argc, char **argv)
     P1("1 == %d", jx_list_check(list));
     P1("0 == %d", jx_hash_check(list));
 
-    P1("0 == %d", jx_list_size(list));
+    P1("0 == %d", (int)jx_list_size(list));
 
     P1("0 == %d", jx_list_append(list, jx_ob_from_int(9)));
-    P1("1 == %d", jx_list_size(list));
-    P1("9 == %d", jx_ob_as_int(jx_list_borrow(list, 0)));
+    P1("1 == %d", (int)jx_list_size(list));
+    P1("9 == %d", (int)jx_ob_as_int(jx_list_borrow(list, 0)));
     P1("0 == %d", jx_list_append(list, jx_ob_from_int(8)));
-    P1("2 == %d", jx_list_size(list));
-    P1("8 == %d", jx_ob_as_int(jx_list_borrow(list, 1)));
+    P1("2 == %d", (int)jx_list_size(list));
+    P1("8 == %d", (int)jx_ob_as_int(jx_list_borrow(list, 1)));
 
     P1("0 == %d", jx_list_append(list, jx_ob_from_float(2.0F)));
-    P1("3 == %d", jx_list_size(list));
-    P1("9 == %d", jx_ob_as_int(jx_list_borrow(list, 0)));
-    P1("8 == %d", jx_ob_as_int(jx_list_borrow(list, 1)));
+    P1("3 == %d", (int)jx_list_size(list));
+    P1("9 == %d", (int)jx_ob_as_int(jx_list_borrow(list, 0)));
+    P1("8 == %d", (int)jx_ob_as_int(jx_list_borrow(list, 1)));
     P1("2.0 == %f", jx_ob_as_float(jx_list_borrow(list, 2)));
 
     P1("0 == %d", jx_ob_free(list));
@@ -75,8 +75,8 @@ int main(int argc, char **argv)
     }
     {
       jx_ob list = jx_list_new_with_int_vla(&vla);
-      P1("4 == %d", jx_list_size(list));
-      P1("3 == %d", jx_ob_as_int(jx_list_borrow(list, 3)));
+      P1("4 == %d", (int)jx_list_size(list));
+      P1("3 == %d", (int)jx_ob_as_int(jx_list_borrow(list, 3)));
       P1("0 == %d", jx_ob_free(list));
     }
   }
@@ -88,7 +88,7 @@ int main(int argc, char **argv)
     }
     {
       jx_ob list = jx_list_new_with_float_vla(&vla);
-      P1("4 == %d", jx_list_size(list));
+      P1("4 == %d", (int)jx_list_size(list));
       P1("0.3 == %f", jx_ob_as_float(jx_list_borrow(list, 3)));
       P1("0 == %d", jx_ob_free(list));
     }
